Extracted Show ProcessList code to separate class
